Fill in HTML content
the html is organized into three main parts i.e. header, main, and
footer to make it easier to display the entire body as flex with
flex-direction: column;
elements to be styled the same way share atleast one class e.g. the
container class.
used the blockquote element in the quote-section to test if it offers
more inherent styling that the p element